requirements.txt
coreapi==2.3.3
Django==3.0.6
django-admin-sortable2==0.7.6
django-environ==0.4.5
django-heroku==0.3.1
django-import-export==2.1.0
django-mathfilters==1.0.0
django-oauth-toolkit==1.3.2
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
django-simple-history==2.10.0
django-statici18n==1.9.0
django-storages[google]==1.9.1
django-xforwardedfor-middleware==2.0
geoip2==3.0.0
gunicorn==20.0.4
phonenumbers==8.12.4
pillow==7.1.2
